Store data as char in Fields, and a string for the datatype
This allows to store integer and FP fields in the same kind
of structure, simplifying the interfaces. The user only needs
to provide the correct data type when extracting the data
(either raw or as a view). The class will check that the provided
type matches what was used at the creation of the field.